Add Flutter Module type on project open (#316).
* adds flutter module type (if missing)
* reloads project
* runs `flutter packages get`

Fixes: #316.
