+++ src/io/flutter/project/FlutterProjectOpenProcessor.java
@@ -10,7 +10,9 @@ import com.intellij.notification.Notification;
 import com.intellij.notification.NotificationListener;
 import com.intellij.notification.NotificationType;
 import com.intellij.notification.Notifications;
+import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.extensions.Extensions;
+import com.intellij.openapi.module.Module;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.projectImport.ProjectOpenProcessor;
@@ -23,16 +25,21 @@ import io.flutter.actions.FlutterPackagesUpgradeAction;
 import io.flutter.actions.FlutterSdkAction;
 import io.flutter.sdk.FlutterSdk;
 import io.flutter.sdk.FlutterSdkUtil;
+import io.flutter.utils.FlutterModuleUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.swing.*;
 import javax.swing.event.HyperlinkEvent;
+import java.text.MessageFormat;
 import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
 
 
 public class FlutterProjectOpenProcessor extends ProjectOpenProcessor {
+
+  private static final Logger LOG = Logger.getInstance(FlutterProjectOpenProcessor.class.getName());
 
   private static void doPerform(@NotNull FlutterSdkAction action, @NotNull Project project) throws ExecutionException {
     final FlutterSdk sdk = FlutterSdk.getFlutterSdk(project);
@@ -87,18 +94,27 @@ public class FlutterProjectOpenProcessor extends ProjectOpenProcessor {
       return;
     }
 
-    // TODO(pq): handle adding Flutter module type.
-    if (!FlutterSdkUtil.hasFlutterModule(project)) {
-      return;
+    if (!FlutterModuleUtils.hasFlutterModule(project)) {
+      if (FlutterModuleUtils.usesFlutter(project)) {
+        final List<Module> modules = FlutterModuleUtils.findModulesWithFlutterContents(project);
+        if (modules.isEmpty()) {
+          LOG.warn(MessageFormat.format("No module found for {0}", project.getName()));
+          return;
+        } else if (modules.size() > 1){
+          LOG.warn(MessageFormat.format("{0} contains {1} modules.", project.getName(), modules.size()));
+        }
+
+        FlutterModuleUtils.setFlutterModuleAndReload(modules.get(0), project);
+      }
     }
 
     try {
-      final VirtualFile packagesFile = FlutterSdkUtil.findPackagesFileFrom(project, null);
+      final VirtualFile packagesFile = FlutterModuleUtils.findPackagesFileFrom(project, null);
       if (!FlutterUtils.exists(packagesFile)) {
         doPerform(new FlutterPackagesGetAction(), project);
       }
       else {
-        final VirtualFile pubspecFile = FlutterSdkUtil.findPubspecFrom(project, null);
+        final VirtualFile pubspecFile = FlutterModuleUtils.findPubspecFrom(project, null);
         if (FlutterUtils.exists(pubspecFile) && pubspecFile.getTimeStamp() > packagesFile.getTimeStamp()) {
           Notifications.Bus.notify(new PackagesOutOfDateNotification(project));
         }
